+++ arch/arm/rpi2/source/ArchBoardSpecific.cpp
@@ -35,7 +35,7 @@ pointer ArchBoardSpecific::getVESAConsoleLFBPtr()
   return framebuffer;
 }
 
-uint32 ArchBoardSpecific::getUsableMemoryRegion(uint32 region __attribute__((unused)), pointer &start_address, pointer &end_address, uint32 &type)
+size_t ArchBoardSpecific::getUsableMemoryRegion(size_t region __attribute__((unused)), pointer &start_address, pointer &end_address, size_t &type)
 {
   start_address = 0;
   end_address = ((PHYSICAL_MEMORY_AVAILABLE - ArchCommon::getVESAConsoleWidth() * ArchCommon::getVESAConsoleHeight() * ArchCommon::getVESAConsoleBitsPerPixel() / 8) & ~0xFFF);
